+++ server-php/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.0 - 2018-05-06
 - Disallow to check private IP/host (can be disabled).
+- Add regex rules support for whitelisted domains.
 
 ## 1.1.0 - 2018-01-08
 - Code rewrite.